+++ packages/dice-physics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dice-o-rolla/dice-physics
 
+## 0.4.0
+
+### Patch Changes
+
+- @dice-o-rolla/dice-core@0.4.0
+
 ## 0.3.1
 
 ### Patch Changes